refactor(tasks): split task recurrence selector into modular components
The large TaskRecurrenceSelector widget has been broken down into smaller,
focused components:
- RecurrenceEndConditionSelector for handling end condition logic
- RecurrenceWeekdaySelector for weekday selection UI
- Main TaskRecurrenceSelector containing core logic

This improves code maintainability, readability, and testability by
separating concerns within the recurrence configuration UI.
